+++ client/src/api/block_builder/input_selection/helpers.rs
@@ -7,7 +7,7 @@ use iota_types::block::{
     address::{Address, AliasAddress, Ed25519Address, NftAddress},
     output::{
         unlock_condition::{AddressUnlockCondition, StorageDepositReturnUnlockCondition},
-        BasicOutputBuilder, NativeTokens, NativeTokensBuilder, Output, OutputId, Rent, RentStructure, UnlockCondition,
+        BasicOutputBuilder, NativeTokens, NativeTokensBuilder, Output, Rent, RentStructure, UnlockCondition,
     },
 };
 
@@ -166,88 +166,4 @@ pub(crate) fn sort_input_signing_data(inputs: Vec<InputSigningData>) -> crate::R
     }
 
     Ok(sorted_inputs)
-}
-
-// Check if an address is required for unlocking an output in any unlock condition
-// Also returns true if the output is an alias or foundry address and the address to search for matches this one
-// If the output id is provided and the alias/nft address is null, we will calculate the correct one.
-pub(crate) fn output_contains_address(
-    output: &Output,
-    output_id: Option<OutputId>,
-    required_address: &Address,
-    current_time: u32,
-) -> bool {
-    // Check alias and nft addresses
-    match output {
-        Output::Alias(alias_output) => {
-            let alias_id = match output_id {
-                Some(output_id) => alias_output.alias_id().or_from_output_id(output_id),
-                None => *alias_output.alias_id(),
-            };
-            if *required_address == Address::Alias(AliasAddress::new(alias_id)) {
-                return true;
-            }
-        }
-        Output::Nft(nft_output) => {
-            let nft_id = match output_id {
-                Some(output_id) => nft_output.nft_id().or_from_output_id(output_id),
-                None => *nft_output.nft_id(),
-            };
-            if *required_address == Address::Nft(NftAddress::new(nft_id)) {
-                return true;
-            }
-        }
-        _ => {}
-    }
-
-    // Check unlock conditions
-    if let Some(unlock_conditions) = output.unlock_conditions() {
-        if let Some(address_unlock_condition) = unlock_conditions.address() {
-            if required_address == unlock_conditions.locked_address(address_unlock_condition.address(), current_time) {
-                return true;
-            }
-        }
-        if let Some(state_controller_unlock_condition) = unlock_conditions.state_controller_address() {
-            if required_address == state_controller_unlock_condition.address() {
-                return true;
-            }
-        }
-        if let Some(governor_controller_unlock_condition) = unlock_conditions.governor_address() {
-            if required_address == governor_controller_unlock_condition.address() {
-                return true;
-            }
-        }
-        if let Some(immutable_alias_address_unlock_condition) = unlock_conditions.immutable_alias_address() {
-            if required_address == immutable_alias_address_unlock_condition.address() {
-                return true;
-            }
-        }
-    }
-
-    false
-}
-
-pub(crate) fn is_basic_output_address_unlockable(output: &Output, address: &Address, current_time: u32) -> bool {
-    match output {
-        Output::Basic(_) => {
-            if let Some(unlock_conditions) = output.unlock_conditions() {
-                if unlock_conditions.is_time_locked(current_time) {
-                    return false;
-                }
-
-                if let Some(expiration) = unlock_conditions.expiration() {
-                    if let Some(expiration_address) = expiration.return_address_expired(current_time) {
-                        return expiration_address == address;
-                    }
-                }
-
-                // PANIC: safe to unwrap as basic outputs always have an address.
-                unlock_conditions.address().unwrap().address() == address
-            } else {
-                // Should not happen anyway as there should always be at least the address unlock condition.
-                false
-            }
-        }
-        _ => false,
-    }
 }