+++ client/src/api/block_builder/input_selection/sender_issuer.rs
@@ -12,11 +12,7 @@ use iota_types::block::{
 };
 
 use crate::{
-    api::{
-        address::search_address,
-        input_selection::{helpers::is_basic_output_address_unlockable, output_contains_address},
-        ClientBlockBuilder,
-    },
+    api::{address::search_address, ClientBlockBuilder},
     constants::HD_WALLET_TYPE,
     secret::types::{InputSigningData, OutputMetadata},
     Error, Result,
@@ -34,29 +30,39 @@ impl<'a> ClientBlockBuilder<'a> {
         let current_time = self.client.get_time_checked()?;
         let token_supply = self.client.get_token_supply()?;
 
-        let all_required_addresses = get_required_addresses_for_sender_and_issuer(&[], &self.outputs, current_time)?;
+        let required_sender_or_issuer_addresses =
+            get_required_addresses_for_sender_and_issuer(&[], &self.outputs, current_time)?;
 
-        for address in all_required_addresses {
-            match address {
+        for sender_or_issuer_address in required_sender_or_issuer_addresses {
+            match sender_or_issuer_address {
                 Address::Ed25519(_) => {
+                    // Check if the address is derived from the seed
                     let (address_index, internal) = search_address(
                         self.secret_manager.ok_or(Error::MissingParameter("secret manager"))?,
                         &bech32_hrp,
                         self.coin_type,
                         self.account_index,
                         self.input_range.clone(),
-                        &address,
+                        &sender_or_issuer_address,
                     )
                     .await?;
-                    // If it didn't return with an error, then the address was found.
-
-                    let address_outputs = self.address_outputs(address.to_bech32(&bech32_hrp)).await?;
+                    let address_outputs = self
+                        .basic_address_outputs(sender_or_issuer_address.to_bech32(&bech32_hrp))
+                        .await?;
 
                     let mut found_output = false;
                     for output_response in address_outputs {
                         let output = Output::try_from_dto(&output_response.output, token_supply)?;
 
-                        if is_basic_output_address_unlockable(&output, &address, current_time) {
+                        // We can ignore the unlocked_alias_or_nft_address, since we only requested basic outputs
+                        let (required_unlock_address, _unlocked_alias_or_nft_address) = output
+                            .required_and_unlocked_address(
+                                current_time,
+                                output_response.metadata.output_id()?,
+                                false,
+                            )?;
+
+                        if required_unlock_address == sender_or_issuer_address {
                             required_inputs.push(InputSigningData {
                                 output,
                                 output_metadata: OutputMetadata::try_from(&output_response.metadata)?,
@@ -67,10 +73,8 @@ impl<'a> ClientBlockBuilder<'a> {
                                     internal as u32,
                                     address_index,
                                 ])),
-                                bech32_address: address.to_bech32(&bech32_hrp),
+                                bech32_address: sender_or_issuer_address.to_bech32(&bech32_hrp),
                             });
-                            // we want to include all outputs, because another output might be better balance wise,
-                            // but will not unlock the address we need
                             found_output = true;
                             break;
                         }
@@ -222,112 +226,58 @@ pub(crate) fn select_inputs_for_sender_and_issuer<'a>(
 ) -> Result<()> {
     log::debug!("[select_inputs_for_sender_and_issuer]");
 
-    let all_required_addresses = get_required_addresses_for_sender_and_issuer(selected_inputs, outputs, current_time)?;
-    'addresses_loop: for address in all_required_addresses {
-        match address {
-            Address::Ed25519(_) => {
-                // first check already selected outputs
-                for input_signing_data in selected_inputs.iter() {
-                    if output_contains_address(
-                        &input_signing_data.output,
-                        Some(input_signing_data.output_id()?),
-                        &address,
-                        current_time,
-                    ) {
-                        continue 'addresses_loop;
-                    }
-                }
+    let required_sender_or_issuer_addresses =
+        get_required_addresses_for_sender_and_issuer(selected_inputs, outputs, current_time)?;
+    'addresses_loop: for required_address in required_sender_or_issuer_addresses {
+        // first check already selected outputs
+        for input_signing_data in selected_inputs.iter() {
+            // Default to `true`, since it will be a state transition if we add it here
+            let alias_state_transition = alias_state_transition(input_signing_data, outputs)?.unwrap_or(true);
+            let (required_unlock_address, unlocked_alias_or_nft_address) = input_signing_data
+                .output
+                .required_and_unlocked_address(current_time, input_signing_data.output_id()?, alias_state_transition)?;
 
-                // if not found, check currently not selected outputs
-                for input_signing_data in inputs.clone() {
-                    // Skip already added inputs
-                    let output_id = input_signing_data.output_id()?;
-                    if selected_inputs_output_ids.contains(&output_id) {
-                        continue;
-                    }
-                    if output_contains_address(&input_signing_data.output, Some(output_id), &address, current_time) {
-                        selected_inputs.push(input_signing_data.clone());
-                        selected_inputs_output_ids.insert(output_id);
-                        // break when we added the required output to the selected_inputs
-                        continue 'addresses_loop;
-                    }
-                }
-
-                return Err(Error::MissingInput(format!(
-                    "missing input with {address:?} for sender or issuer feature"
-                )));
+            if required_unlock_address == required_address {
+                continue 'addresses_loop;
             }
-            Address::Alias(alias_address) => {
-                // Check if output is alias address.
-                let alias_id = alias_address.alias_id();
-
-                // check if already selected.
-                for selected_input in selected_inputs.iter() {
-                    if let Output::Alias(alias_output) = &selected_input.output {
-                        if alias_id == &alias_output.alias_id().or_from_output_id(selected_input.output_id()?) {
-                            continue 'addresses_loop;
-                        }
-                    }
+            if let Some(unlocked_alias_or_nft_address) = unlocked_alias_or_nft_address {
+                if unlocked_alias_or_nft_address == required_address {
+                    continue 'addresses_loop;
                 }
-
-                // if not found, check currently not selected outputs
-                for input_signing_data in inputs.clone() {
-                    // Skip already added inputs
-                    let output_id = input_signing_data.output_id()?;
-                    if selected_inputs_output_ids.contains(&output_id) {
-                        continue;
-                    }
-                    if let Output::Alias(alias_output) = &input_signing_data.output {
-                        if alias_id
-                            == &alias_output
-                                .alias_id()
-                                .or_from_output_id(input_signing_data.output_id()?)
-                        {
-                            selected_inputs.push(input_signing_data.clone());
-                            selected_inputs_output_ids.insert(output_id);
-                            // break when we added the required output to the selected_inputs
-                            continue 'addresses_loop;
-                        }
-                    }
-                }
-                return Err(Error::MissingInput(format!(
-                    "missing alias {alias_id} for sender or issuer feature"
-                )));
-            }
-            Address::Nft(nft_address) => {
-                // Check if output is nft address.
-                let nft_id = nft_address.nft_id();
-
-                // check if already selected.
-                for selected_input in selected_inputs.iter() {
-                    if let Output::Nft(nft_output) = &selected_input.output {
-                        if nft_id == &nft_output.nft_id().or_from_output_id(selected_input.output_id()?) {
-                            continue 'addresses_loop;
-                        }
-                    }
-                }
-
-                // if not found, check currently not selected outputs
-                for input_signing_data in inputs.clone() {
-                    // Skip already added inputs
-                    let output_id = input_signing_data.output_id()?;
-                    if selected_inputs_output_ids.contains(&output_id) {
-                        continue;
-                    }
-                    if let Output::Nft(nft_output) = &input_signing_data.output {
-                        if nft_id == &nft_output.nft_id().or_from_output_id(input_signing_data.output_id()?) {
-                            selected_inputs.push(input_signing_data.clone());
-                            selected_inputs_output_ids.insert(output_id);
-                            // break when we added the required output to the selected_inputs
-                            continue 'addresses_loop;
-                        }
-                    }
-                }
-                return Err(Error::MissingInput(format!(
-                    "missing nft {nft_id} for sender or issuer feature"
-                )));
             }
         }
+
+        // if not found, check currently not selected outputs
+        for input_signing_data in inputs.clone() {
+            // Skip already added inputs
+            let output_id = input_signing_data.output_id()?;
+            if selected_inputs_output_ids.contains(&output_id) {
+                continue;
+            }
+
+            // Default to `true`, since it will be a state transition if we add it here
+            let alias_state_transition = alias_state_transition(input_signing_data, outputs)?.unwrap_or(true);
+            let (required_unlock_address, unlocked_alias_or_nft_address) = input_signing_data
+                .output
+                .required_and_unlocked_address(current_time, output_id, alias_state_transition)?;
+
+            if required_unlock_address == required_address {
+                selected_inputs.push(input_signing_data.clone());
+                selected_inputs_output_ids.insert(output_id);
+                continue 'addresses_loop;
+            }
+            if let Some(unlocked_alias_or_nft_address) = unlocked_alias_or_nft_address {
+                if unlocked_alias_or_nft_address == required_address {
+                    selected_inputs.push(input_signing_data.clone());
+                    selected_inputs_output_ids.insert(output_id);
+                    continue 'addresses_loop;
+                }
+            }
+        }
+
+        return Err(Error::MissingInput(format!(
+            "missing input with {required_address:?} for sender or issuer feature"
+        )));
     }
 
     Ok(())
@@ -341,21 +291,30 @@ fn get_required_addresses_for_sender_and_issuer(
 ) -> crate::Result<HashSet<Address>> {
     log::debug!("[get_required_addresses_for_sender_and_issuer]");
 
-    let mut all_required_addresses = HashSet::new();
+    // Addresses in the inputs that will be unlocked in the transaction
+    let mut unlocked_addresses = HashSet::new();
+    for input_signing_data in selected_inputs {
+        let alias_state_transition = alias_state_transition(input_signing_data, outputs)?;
+        let (required_unlock_address, unlocked_alias_or_nft_address) =
+            input_signing_data.output.required_and_unlocked_address(
+                current_time,
+                input_signing_data.output_id()?,
+                alias_state_transition.unwrap_or(false),
+            )?;
+        unlocked_addresses.insert(required_unlock_address);
+        if let Some(unlocked_alias_or_nft_address) = unlocked_alias_or_nft_address {
+            unlocked_addresses.insert(unlocked_alias_or_nft_address);
+        }
+    }
+
+    let mut required_sender_or_issuer_addresses = HashSet::new();
 
     for output in outputs {
         if let Some(sender_feature) = output.features().and_then(Features::sender) {
-            if !all_required_addresses.contains(sender_feature.address()) {
+            if !required_sender_or_issuer_addresses.contains(sender_feature.address()) {
                 // Only add if not already present in the selected inputs.
-                if !selected_inputs.iter().any(|input_data| {
-                    output_contains_address(
-                        &input_data.output,
-                        Some(input_data.output_id().expect("invalid output id in input signing data")),
-                        sender_feature.address(),
-                        current_time,
-                    )
-                }) {
-                    all_required_addresses.insert(*sender_feature.address());
+                if !unlocked_addresses.contains(sender_feature.address()) {
+                    required_sender_or_issuer_addresses.insert(*sender_feature.address());
                 }
             }
         }
@@ -368,22 +327,49 @@ fn get_required_addresses_for_sender_and_issuer(
         };
         if utxo_chain_creation {
             if let Some(issuer_feature) = output.immutable_features().and_then(Features::issuer) {
-                if !all_required_addresses.contains(issuer_feature.address()) {
+                if !required_sender_or_issuer_addresses.contains(issuer_feature.address()) {
                     // Only add if not already present in the selected inputs.
-                    if !selected_inputs.iter().any(|input_data| {
-                        output_contains_address(
-                            &input_data.output,
-                            Some(input_data.output_id().expect("invalid output id in input signing data")),
-                            issuer_feature.address(),
-                            current_time,
-                        )
-                    }) {
-                        all_required_addresses.insert(*issuer_feature.address());
+                    if !unlocked_addresses.contains(issuer_feature.address()) {
+                        required_sender_or_issuer_addresses.insert(*issuer_feature.address());
                     }
                 }
             }
         }
     }
 
-    Ok(all_required_addresses)
+    Ok(required_sender_or_issuer_addresses)
+}
+
+// Returns if alias transition is a state transition with the provided outputs for a given input.
+pub(crate) fn alias_state_transition(
+    input_signing_data: &InputSigningData,
+    outputs: &[Output],
+) -> Result<Option<bool>> {
+    Ok(if let Output::Alias(alias_input) = &input_signing_data.output {
+        let alias_id = alias_input
+            .alias_id()
+            .or_from_output_id(input_signing_data.output_id()?);
+        // Check if alias exists in the outputs and get the required transition type
+        outputs
+            .iter()
+            .find_map(|o| {
+                if let Output::Alias(alias_output) = o {
+                    if *alias_output.alias_id() == alias_id {
+                        if alias_output.state_index() == alias_input.state_index() {
+                            Some(Some(false))
+                        } else {
+                            Some(Some(true))
+                        }
+                    } else {
+                        None
+                    }
+                } else {
+                    None
+                }
+                // if not find in the outputs, the alias gets burned which is a governance transaction
+            })
+            .unwrap_or(None)
+    } else {
+        None
+    })
 }