+++ types/src/block/output/output_id.rs
@@ -43,6 +43,15 @@ impl OutputId {
         self.index.get()
     }
 
+    /// Creates a null [`OutputId`].
+    pub fn null() -> Self {
+        Self {
+            transaction_id: TransactionId::null(),
+            // Unwrap is fine because index is already known and valid.
+            index: 0u16.try_into().unwrap(),
+        }
+    }
+
     /// Splits an [`OutputId`] into its [`TransactionId`] and index.
     #[inline(always)]
     pub fn split(self) -> (TransactionId, u16) {