+++ types/CHANGELOG.md
@@ -30,6 +30,7 @@ First release based on `bee-api-types` and `bee-block`.
 - `required_and_unlocked_address()` to `Output`;
 - `output_id()` to `OutputMetadataResponse`;
 - Independent `From`/`TryFrom` implementations for all unlock conditions DTOs;
+- `OutputId::null`;
 
 ### Changed